Set up Provider and Consumer
Set up Provider and Consumer but still not working as desired
