+++ src/components/header.js
@@ -1,12 +1,10 @@
 import React from 'react'
 import Nav from './nav'
-import Toggle from './toggle'
 
 const Header = () => {
     return (
         <>
             <Nav />
-            <Toggle />
         </>
     )
 }
